+++ src/Services/Sale/Property/Result/PropertyFieldsByTypeResult.php
@@ -1,0 +1,33 @@
+<?php
+
+/**
+ * This file is part of the bitrix24-php-sdk package.
+ *
+ * © Sally Fancen <vadimsallee@gmail.com>
+ *
+ * For the full copyright and license information, please view the MIT-LICENSE.txt
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
+namespace Bitrix24\SDK\Services\Sale\Property\Result;
+
+use Bitrix24\SDK\Core\Exceptions\BaseException;
+use Bitrix24\SDK\Core\Result\AbstractResult;
+
+/**
+ * Class PropertyFieldsByTypeResult
+ *
+ * Result for sale.property.getFieldsByType
+ */
+class PropertyFieldsByTypeResult extends AbstractResult
+{
+    /**
+     * @throws BaseException
+     */
+    public function getFieldsDescription(): array
+    {
+        return $this->getCoreResponse()->getResponseData()->getResult()['property'];
+    }
+}